pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aesthetic-scorer-mcp"
version = "0.1.0"
description = "MCP server for image aesthetic scoring using rsinema/aesthetic-scorer model"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]
